+++ dev/sg/sg_update.go
@@ -59,7 +59,7 @@ func updateToPrebuiltSG(ctx context.Context) (string, error) {
 	if err != nil {
 		return "", err
 	}
-	if err := download.Exeuctable(downloadURL, currentExecPath); err != nil {
+	if err := download.Executable(downloadURL, currentExecPath); err != nil {
 		return "", err
 	}
 	return currentExecPath, nil