+++ enterprise/cmd/worker/internal/codeintel/indexing_config.go
@@ -3,7 +3,6 @@ package codeintel
 import (
 	"time"
 
-	"github.com/sourcegraph/sourcegraph/internal/codeintel/autoindexing/enqueuer"
 	"github.com/sourcegraph/sourcegraph/internal/env"
 	"github.com/sourcegraph/sourcegraph/lib/errors"
 )
@@ -11,7 +10,6 @@ import (
 type indexingConfig struct {
 	env.BaseConfig
 
-	AutoIndexEnqueuerConfig                *enqueuer.Config
 	AutoIndexingTaskInterval               time.Duration
 	RepositoryProcessDelay                 time.Duration
 	RepositoryBatchSize                    int
@@ -23,10 +21,6 @@ type indexingConfig struct {
 var indexingConfigInst = &indexingConfig{}
 
 func (c *indexingConfig) Load() {
-	enqueuerConfig := &enqueuer.Config{}
-	enqueuerConfig.Load()
-	indexingConfigInst.AutoIndexEnqueuerConfig = enqueuerConfig
-
 	c.AutoIndexingTaskInterval = c.GetInterval("PRECISE_CODE_INTEL_AUTO_INDEXING_TASK_INTERVAL", "10m", "The frequency with which to run periodic codeintel auto-indexing tasks.")
 	c.RepositoryProcessDelay = c.GetInterval("PRECISE_CODE_INTEL_AUTO_INDEXING_REPOSITORY_PROCESS_DELAY", "24h", "The minimum frequency that the same repository can be considered for auto-index scheduling.")
 	c.RepositoryBatchSize = c.GetInt("PRECISE_CODE_INTEL_AUTO_INDEXING_REPOSITORY_BATCH_SIZE", "100", "The number of repositories to consider for auto-indexing scheduling at a time.")
@@ -38,6 +32,5 @@ func (c *indexingConfig) Load() {
 func (c *indexingConfig) Validate() error {
 	var errs error
 	errs = errors.Append(errs, c.BaseConfig.Validate())
-	errs = errors.Append(errs, c.AutoIndexEnqueuerConfig.Validate())
 	return errs
 }