+++ enterprise/internal/insights/resolvers/insight_view_resolvers_test.go
@@ -6,18 +6,17 @@ import (
 	"sort"
 	"testing"
 
-	"github.com/sourcegraph/sourcegraph/cmd/frontend/graphqlbackend"
-	"github.com/sourcegraph/sourcegraph/internal/actor"
-	"github.com/sourcegraph/sourcegraph/internal/timeutil"
-	internalTypes "github.com/sourcegraph/sourcegraph/internal/types"
-
-	"github.com/sourcegraph/sourcegraph/enterprise/internal/insights/store"
-	"github.com/sourcegraph/sourcegraph/internal/database/basestore"
-	"github.com/sourcegraph/sourcegraph/internal/database/dbtest"
-
 	"github.com/google/go-cmp/cmp"
 
+	"github.com/sourcegraph/sourcegraph/cmd/frontend/graphqlbackend"
+	"github.com/sourcegraph/sourcegraph/enterprise/internal/insights/store"
 	"github.com/sourcegraph/sourcegraph/enterprise/internal/insights/types"
+	"github.com/sourcegraph/sourcegraph/internal/actor"
+	"github.com/sourcegraph/sourcegraph/internal/database"
+	"github.com/sourcegraph/sourcegraph/internal/database/basestore"
+	"github.com/sourcegraph/sourcegraph/internal/database/dbtest"
+	"github.com/sourcegraph/sourcegraph/internal/timeutil"
+	internalTypes "github.com/sourcegraph/sourcegraph/internal/types"
 )
 
 func addrStr(input string) *string {
@@ -167,7 +166,7 @@ func TestFrozenInsightDataSeriesResolver(t *testing.T) {
 	})
 	t.Run("insight_is_not_frozen_returns_real_resolvers", func(t *testing.T) {
 		insightsDB := dbtest.NewInsightsDB(t)
-		postgres := dbtest.NewDB(t)
+		postgres := database.NewDB(dbtest.NewDB(t))
 		permStore := store.NewInsightPermissionStore(postgres)
 		clock := timeutil.Now
 		timeseriesStore := store.NewWithClock(insightsDB, permStore, clock)
@@ -231,7 +230,7 @@ func TestInsightViewDashboardConnections(t *testing.T) {
 	ctx := actor.WithActor(context.Background(), a)
 
 	insightsDB := dbtest.NewInsightsDB(t)
-	postgresDB := dbtest.NewDB(t)
+	postgresDB := database.NewDB(dbtest.NewDB(t))
 	base := baseInsightResolver{
 		insightStore:   store.NewInsightStore(insightsDB),
 		dashboardStore: store.NewDashboardStore(insightsDB),