+++ internal/codeintel/autoindexing/gen.go
@@ -1,3 +1,3 @@
 package autoindexing
 
-//go:generate ../../../dev/mockgen.sh github.com/sourcegraph/sourcegraph/internal/codeintel/autoindexing/internal/store -i Store -o mock_iface_test.go
+//go:generate ../../../dev/mockgen.sh github.com/sourcegraph/sourcegraph/internal/codeintel/autoindexing -i DBStore -i GitServerClient -i RepoUpdaterClient -i InferenceService -o mock_iface_test.go